Use a view setting to control deletion
Include in the settings added to the view created when a file is
unzipped a setting to determine if the file should be deleted when its
closed or not.

For now this defaults to being turned on because the files that its set
for are temporary files created to make it possible for the user to edit
the file, so they should not hang around.

+++ GZipper/gzipper.py
@@ -73,6 +73,7 @@ class ReopenAsGzipCommand(sublime_plugin.TextCommand):
             # tracking.
             gzView = self.view.window().open_file(new_name)
             gzView.settings().set("_gzip_name", org_name)
+            gzView.settings().set("_gzip_delete", True)
 
             # Flag the view as a gzipped file
             gzView.set_status("gzipper", "[gzipped file]")
@@ -107,7 +108,8 @@ class GzipFileListener(sublime_plugin.ViewEventListener):
         return settings.has("_gzip_name")
 
     def on_close(self):
-        os.remove(self.view.file_name())
+        if self.view.settings().get("_gzip_delete", False):
+            os.remove(self.view.file_name())
 
     def on_post_save(self):
         org_filename = self.view.settings().get("_gzip_name")
